refactor: extract common build entrypoint config
